+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Fixed
 
+- fix dot notation handling in toObject and update
 - updated changelog github action
 
 ## 1.3.2